Changed mapping for numbers to long and double.

+++ structs4java-maven-plugin-test/src/main/java/org/structs4java/example/tests/InterfaceB.java
@@ -2,5 +2,5 @@ package org.structs4java.example.tests;
 
 public interface InterfaceB {
 
-	public int getB();
+	public long getB();
 }
